% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipboardjs-package.R
\docType{package}
\name{clipboardjs-package}
\alias{clipboardjs}
\alias{clipboardjs-package}
\title{clipboardjs: A Wrapper for Clipboard.js}
\description{
Wraps the clipboard.js wrapper to make it easy to add "copy to clipboard" to Shiny apps.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ndiquattro/clipboardjs}
  \item Report bugs at \url{https://github.com/ndiquattro/clipboardjs/issues}
}

}
\author{
\strong{Maintainer}: Nick DiQuattro \email{ndiquattro@gmail.com}

}
\keyword{internal}
