% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copyButton.R
\name{copyButton}
\alias{copyButton}
\title{Copy button}
\usage{
copyButton(id, target_id = NULL, text = NULL, cut = FALSE, label = "Copy", ...)
}
\arguments{
\item{id}{The \code{inputId} for the underlying \code{actionButton}.}

\item{target_id}{The id of the UI output to copy.}

\item{text}{Value to copy to the clipboard when the button is clicked.}

\item{cut}{Cut text to clipboard.}

\item{label}{Label of the button.}

\item{...}{Other arguments passed to \code{\link[shiny:actionButton]{shiny::actionButton()}}.}
}
\value{
An \code{actionButton} enabled to copy to clipboard.
}
\description{
Create a copy button that will copy the contents of a target UI element or a
provided value to the clipboard.
}
\details{
To update the value that will be copied from the server see
\code{\link[=updateCopyText]{updateCopyText()}}.
}
\examples{
if (interactive()) {
  ui <- fluidPage(
    textOutput("test_text"),

    # Copy other output
    copyButton("cp1", target_id = "test_text", icon = icon("clipboard")),

    # Copy a set value
    copyButton("cp2", text = "This is the text you want!", label = "Copy Text")
  )

  server <- function(input, output, session) {
    output$test_text <- renderText("The text we are to copy!")
  }

  shinyApp(ui, server)
}
}
