% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pasteable.R
\name{pasteable}
\alias{pasteable}
\title{Make R objects pasteable}
\usage{
pasteable(x, target = c("sheet", "chat"))
}
\arguments{
\item{x}{A \code{data.frame} or \code{list}.}

\item{target}{Where you will paste the content.}
}
\value{
A length 1 character vector.
}
\description{
Make R objects pasteable
}
\examples{
# Meaningful row names are kept
pasteable(mtcars)

# Row names that are a sequence are dropped
pasteable(iris)

# Fixed width output for sharing on slack or other markdown destination
pasteable(mtcars, target = "chat")

# Lists are simply converted
pasteable(list(123, "abc", cool = "R"), target = "chat")
}
