% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateCopyText.R
\name{updateCopyText}
\alias{updateCopyText}
\title{updateCopyText}
\usage{
updateCopyText(id, text, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{id}{The id of the button.}

\item{text}{Value to copy to the clipboard when the button is clicked.}

\item{session}{The session object passed to function given to shinyServer.}
}
\description{
Change the text a button will copy to the clipboard from the server.
}
\examples{
if (interactive()) {
  ui <- fluidPage(
    # Copy a set value
    copyButton("cp2", text = "The client side text", label = "Copy Text")
  )

  server <- function(input, output, session) {
    updateCopyText("cp2", text = "The new server side text!")
  }

  shinyApp(ui, server)
}
}
